bring up the main system

+++ src/pref/config/config_definition.rs
@@ -104,7 +104,8 @@ pub struct SysMgmt {
 	#[serde(alias = "inhibitOnBehalf")]
 	pub conduct_inhibit:	bool,
 
-	pub uclamp_max:		u8,
+	#[serde(alias = "uclamp")]
+	pub uclamp_max:		u32,
 
 	#[serde(alias = "deviceAllow")]
 	#[serde(deserialize_with = "deserialise_device_allow")]
